avances

+++ src/Clinica Frba/Clases/Persona.cs
@@ -15,7 +15,7 @@ namespace Clinica_Frba.Clase_Persona
         public string Apellido { get; set; }
         public decimal TipoDocumento { get; set; }
         public decimal NumeroDocumento  { get; set; }
-        public int Telefono { get; set; }
+        public decimal Telefono { get; set; }
         public string Direccion { get; set; }
         public string Mail  { get; set; }
         public char Sexo  { get; set; }
